// src/app/model/business-model.ts
export class Business {
    Id?: number;
    name?: string;
    shortDescription?: string;
    description?: string;
    imageUrl?: string;
    contactNumber?: number;
    website?: string;
    instaLink?: string
    services?: [];
    review?: number;
    reviewCount?: number;
    location?: {
        address: string;
        city: string;
        country: string;
    serviceArea?: []
    }
};